requirements.txt
python-dotenv==1.0.0
python-telegram-bot==13.15
vk-api==11.9.9
redis==4.6.0
